moved load_ips_from_csv definition in luxos.ips
